# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Gym Ledger API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/profile": {
      "get": {
        "summary": "Get user profile",
        "description": "Returns logged-in user profile using authentication token stored in cookies",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Profile data returned successfully"
          },
          "401": {
            "description": "Unauthorized user"
          }
        }
      }
    }
  },
...
}